Jeu tourne en temps réel + meilleure conception

+++ src/model/services/IOperationsSpeeds.java
@@ -2,7 +2,7 @@ package model.services;
 
 public interface IOperationsSpeeds
 {
-	int get(PlayerCommandType type);
-	int get(GuardCommandType type);
-	int hole();
+	long get(PlayerCommandType type);
+	long get(GuardCommandType type);
+	long hole();
 }
